13-HACKATHON-BACKEND-Pawsitive/NestOn/requirements.txt
asgiref==3.9.1
boto3==1.40.13
botocore==1.40.13
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
jmespath==1.0.1
pillow==11.3.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.5
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
